up to middleware

+++ backend/controllers/stuff.js
@@ -64,17 +64,31 @@ exports.createThing = (req, res, next) => {
   };
 
   exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id: req.params.id}).then(
-      () => {
-        res.status(200).json({
-          message: 'Deleted!'
-        });
-      }
-    ).catch(
-      (error) => {
-        res.status(400).json({
-          error: error
-        });
+    Thing.findOne({ _id: req.params.id }).then(
+      (thing) => {
+        if (!thing) {
+          return res.status(404).json({
+            error: new Error('no such thing')
+          });
+        }
+        if (thing.userId !== req.auth.userId) {
+          return res.status(401).json({
+            error: new Error('unauthorized request')
+          });
+        }
+        Thing.deleteOne({_id: req.params.id}).then(
+          () => {
+            res.status(200).json({
+              message: 'Deleted!'
+            });
+          }
+        ).catch(
+          (error) => {
+            res.status(400).json({
+              error: error
+            });
+          }
+        );
       }
     );
   };
